requirements.txt
annotated-types==0.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.18.0
packaging==24.2
platformdirs==4.3.6
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
rsa==4.9
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
Werkzeug==3.1.3
